+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   security settings
 - Add a lti link to classroom in the standalone website
 - Invite link for BBB available in every context
+- Add a check on classroom document file size when uploading content
 
 ## [4.0.0-beta.14] - 2023-01-25